[curve25519] Remove `ElGamalError` from curve25519 crate (#1777)
* remove `ElGamalError` from curve25519 crate

* add `ElGamalError` to zk-token-sdk
