+++ zk-token-sdk/src/zk_token_elgamal/pod/elgamal.rs
@@ -2,9 +2,11 @@
 
 #[cfg(not(target_os = "solana"))]
 use {
-    crate::encryption::elgamal::{self as decoded},
+    crate::{
+        encryption::elgamal::{self as decoded},
+        errors::ElGamalError,
+    },
     curve25519_dalek::ristretto::CompressedRistretto,
-    solana_curve25519::errors::ElGamalError,
 };
 use {
     crate::{
